flattened directory structure
