+++ server/modules/user/user.auth.controllers.js
@@ -1,7 +1,7 @@
-const UserAccess = require("../dataAccess/user.access");
-const security = require("../../../lib/security");
-const mailSender = require("../../../lib/mailSender");
-const ErrorEnum = require("../../../lib/enums/error");
+const UserAccess = require("./dataAccess/user.access");
+const security = require("../../lib/security");
+const mailSender = require("../../lib/mailSender");
+const ErrorEnum = require("../../lib/enums/error");
 
 exports.login = async (req, res) => {
   const { email, password } = req.body;
@@ -32,7 +32,9 @@ exports.login = async (req, res) => {
 
 exports.recoverPassword = async (req, res) => {
   let mailOptions = mailSender.mailOptions;
-  let tempPass = Math.random().toString(36).substring(2, 15);
+  let tempPass = Math.random()
+    .toString(36)
+    .substring(2, 15);
   mailOptions.to = req.body.email;
   mailOptions.subject = "Restablecimiento de contraseña";
   mailOptions.text = `La contraseña temporal de acceso es ${tempPass}`;
@@ -48,15 +50,15 @@ exports.recoverPassword = async (req, res) => {
 
     user.password = await security.hashPassword(tempPass);
 
-    UserAccess.updateUser(user).then(usr => {
-      mailSender.sendMail(mailOptions);
-    }).then(() => {
-      res.send("Correo enviado con éxito");
-    });
-
+    UserAccess.updateUser(user)
+      .then(usr => {
+        mailSender.sendMail(mailOptions);
+      })
+      .then(() => {
+        res.send("Correo enviado con éxito");
+      });
   } catch (err) {
     console.error(err);
     res.status(400).send(err);
   }
-
 };
