+++ server/modules/user/user.auth.routes.js
@@ -6,6 +6,8 @@ const userAuthController = require("../controllers/user.auth.controllers");
 const MODULE_PATH = "/auth-user";
 
 router.route(MODULE_PATH + "/").post(userAuthController.login);
-router.route(MODULE_PATH + "/recoverPassword").post(userAuthController.recoverPassword);
+router
+  .route(MODULE_PATH + "/recoverPassword")
+  .post(userAuthController.recoverPassword);
 
 module.exports = router;
